Payments and Plans

+++ src/main/java/org/knop/budgetKeeper/repository/UserRepository.java
@@ -4,12 +4,11 @@ import org.knop.budgetKeeper.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
 import java.util.Optional;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByLogin(String login);
     Optional<User> findByEmail(String email);
-    Optional<User> findById(Long id);
+    Optional<User> findById(Integer id);
 }
